backend/requirements.txt
# ============================================================
# requirements.txt - Dipendenze Python Backend
# Progetto: Garage Manager (Gestionale Officina)
# Versioni pinnate per produzione
# ============================================================

# Python >= 3.11 richiesto
# Sintassi X | None nei modelli SQLAlchemy richiede Python 3.10+
# Si raccomanda Python 3.11 per supporto LTS

# ------------------------------------------------------------
# FRAMEWORK FASTAPI
# ------------------------------------------------------------
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# ------------------------------------------------------------
# DATABASE - SQLAlchemy 2.0 async + asyncpg
# ------------------------------------------------------------
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ------------------------------------------------------------
# VALIDAZIONE E CONFIGURAZIONE
# ------------------------------------------------------------
pydantic==2.10.3
pydantic-settings==2.6.1

# ------------------------------------------------------------
# UTILITIES
# ------------------------------------------------------------
python-dotenv==1.0.1
email-validator==2.2.0

# ------------------------------------------------------------
# AUTENTICAZIONE JWT
# ------------------------------------------------------------
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0

# ------------------------------------------------------------
# TESTING
# ------------------------------------------------------------
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
